+++ src/dialogs/dialogs.android.ts
@@ -65,7 +65,9 @@ function createAlertDialogBuilder(options?: DialogOptions & MDCAlertControlerOpt
                 : Builder.createViewFromEntry({
                       moduleName: options.view as string
                   });
-
+        if (view.parent) {
+            view.parent._removeView(view)
+        }
         view.cssClasses.add(CSSUtils.MODAL_ROOT_VIEW_CSS_CLASS);
         const modalRootViewCssClasses = CSSUtils.getSystemCssClasses();
         modalRootViewCssClasses.forEach((c) => view.cssClasses.add(c));
@@ -170,9 +172,7 @@ function prepareAndCreateAlertDialog(
             return;
         }
         if (onDoneCalled) {
-            if (toBeCalledBeforeCallback) {
-                toBeCalledBeforeCallback();
-            }
+            toBeCalledBeforeCallback?.();
             return;
         }
         //ensure we hide any keyboard
@@ -191,9 +191,8 @@ function prepareAndCreateAlertDialog(
         if (dialog) {
             dialog.cancel();
         }
-        if (toBeCalledBeforeCallback) {
-            toBeCalledBeforeCallback();
-        }
+        toBeCalledBeforeCallback?.();
+
         callback && callback(result);
     };
     if (!DialogInterface) {
@@ -459,6 +458,9 @@ export function prompt(arg: any): Promise<PromptResult> {
                 }
             }
             stackLayout.addChild(textField);
+            if (options.view instanceof View) {
+                stackLayout.addChild(options.view);
+            }
             options.view = stackLayout;
             const alert = createAlertDialogBuilder(options);
 
@@ -532,6 +534,10 @@ export function login(arg: any): Promise<LoginResult> {
 
             stackLayout.addChild(userNameTextField);
             stackLayout.addChild(passwordTextField);
+
+            if (options.view instanceof View) {
+                stackLayout.addChild(options.view);
+            }
             options.view = stackLayout;
 
             const alert = createAlertDialogBuilder(options);