+++ CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.2.49](https://github.com/nativescript-community/ui-material-component/compare/v7.2.48...v7.2.49) (2024-04-08)
+
+### Bug Fixes
+
+* ensure custom view has no parent ([7d3dd07](https://github.com/nativescript-community/ui-material-component/commit/7d3dd07bb5c56424f45287f907e405732e7f2e41))
+
+## [7.2.48](https://github.com/nativescript-community/ui-material-component/compare/v7.2.47...v7.2.48) (2024-04-08)
+
+### Bug Fixes
+
+* **dialogs:** allow custom view to be added to `prompt/login` (for now under the text fields) ([33eac9d](https://github.com/nativescript-community/ui-material-component/commit/33eac9d44fef3fa04c9f8871e7be2305a7a40c36))
+
+## [7.2.47](https://github.com/nativescript-community/ui-material-component/compare/v7.2.46...v7.2.47) (2024-03-28)
+
+### Bug Fixes
+
+* **android:** working fix for testId in textfield/textview ([b637cac](https://github.com/nativescript-community/ui-material-component/commit/b637cace3a35fb75b89ffc7936a15d29a109e4ca))
+
+## [7.2.46](https://github.com/nativescript-community/ui-material-component/compare/v7.2.45...v7.2.46) (2024-03-28)
+
+### Bug Fixes
+
+* **textview, textfield, android:** accessibility fix ([0f2ce73](https://github.com/nativescript-community/ui-material-component/commit/0f2ce73a0a505ab07f65ce88e1b8263e78271d6b))
+
 ## [7.2.45](https://github.com/nativescript-community/ui-material-component/compare/v7.2.44...v7.2.45) (2024-03-06)
 
 ### Bug Fixes